+++ packages/fans-society-contracts/contracts/tokens/ProjectTokenFactory.sol
@@ -1,15 +1,20 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
-import {Clones} from '@openzeppelin/contracts/proxy/Clones.sol';
+import { Clones } from '@openzeppelin/contracts/proxy/Clones.sol';
 
-import {IProjectTokenFactory} from './interfaces/IProjectTokenFactory.sol';
-import {ProjectTokenERC20} from './ProjectTokenERC20.sol';
+import { IProjectTokenFactory } from './interfaces/IProjectTokenFactory.sol';
+import { ProjectTokenERC20 } from './ProjectTokenERC20.sol';
 
 contract ProjectTokenFactory is IProjectTokenFactory {
 	address private immutable tokenImplementationAddress;
 
-	event TokenCreated(uint256 projectId, address token, string name, string symbol);
+	event TokenCreated(
+		uint256 projectId,
+		address token,
+		string name,
+		string symbol
+	);
 
 	constructor(address _tokenImplementationAddress) {
 		tokenImplementationAddress = _tokenImplementationAddress;
@@ -20,8 +25,8 @@ contract ProjectTokenFactory is IProjectTokenFactory {
 		string memory _name,
 		string memory _symbol,
 		address _amm,
-		uint40 _totalSupply,
-		uint40 _initialSupply
+		uint112 _totalSupply,
+		uint112 _initialSupply
 	) public returns (address) {
 		address token = Clones.clone(tokenImplementationAddress);
 		ProjectTokenERC20(token).initialize(
